yeet 2

+++ src/map/action/action.cpp
@@ -24,7 +24,7 @@
 #include "enums/four_cc.h"
 #include "utils/battleutils.h"
 
-void action_result_t::recordSkillchain(const ActionProcSkillChain effect, const int16_t dmg)
+void action_result_t::recordSkillchain(const ActionProcSkillChain effect, const int32_t dmg) /* SC int32_t definition override */
 {
     if (dmg < 0)
     {
